Refactoring: Flatten Executor and CompilationManager
